convert some js to ts
