+++ frontend/src/js/DotLayer/ActivityCollection.js
@@ -7,7 +7,7 @@
 import * as ColorPalette from "./ColorPalette.js"
 import { Activity } from "./Activity.js"
 import { options } from "./Defaults.js"
-import BitSet from "../BitSet.js"
+import { BitSet } from "../BitSet.ts"
 import { queueTask, nextTask } from "../appUtil.js"
 import { PixelGraphics } from "./PixelGraphics"
 
@@ -145,20 +145,6 @@ export function resetSegMasks() {
     A.resetSegMask()
   }
 }
-
-// // Debug info box
-// let _infoBox
-// const InfoViewer = Control.extend({
-//   onAdd: function () {
-//     _infoBox = DomUtil.create("div")
-//     _infoBox.style.width = "200px"
-//     _infoBox.style.padding = "5px"
-//     _infoBox.style.background = "rgba(100,200,50,0.6)"
-//     _infoBox.style.textAlign = "left"
-//     _infoBox.innerHTML = "Activites infoBox"
-//     return _infoBox
-//   },
-// })
 
 /*
  * Methods for drawing to imageData objects
